Refactored to helper methods and added unit test
