+++ workers/github_worker/github_worker/worker.py
@@ -8,6 +8,7 @@ import requests, time, logging, json, os
 from datetime import datetime
 from sqlalchemy.ext.declarative import declarative_base
 from workers.standard_methods import get_table_values, init_oauths, get_max_id, register_task_completion, register_task_failure, connect_to_broker, update_gh_rate_limit, record_model_process, paginate
+from helpers import Helpers
 
 class GitHubWorker:
     """ Worker that collects data from the Github API and stores it in our database
@@ -984,7 +985,8 @@ class GitHubWorker:
                 j = r.json()
 
                 # Checking contents of requests with what we already have in the db
-                new_events = self.check_duplicates(j, event_table_values, pseudo_key_gh)
+                helper_class = Helpers()
+                new_events = helper_class.check_duplicates(j, event_table_values, pseudo_key_gh)
                 if len(new_events) == 0 and multiple_pages and 'last' in r.links:
                     if i - 1 != int(r.links['last']['url'][-6:].split('=')[1]):
                         logging.info("No more pages with unknown events, breaking from pagination.\n")
@@ -1212,19 +1214,6 @@ class GitHubWorker:
             logging.info("Inserted contributor: " + contributor['login'] + "\n")
             
             return self.find_id_from_login(login)
-            
-    def check_duplicates(self, new_data, table_values, key):
-        need_insertion = []
-        for obj in new_data:
-            if type(obj) == dict:
-                if not table_values.isin([obj[key]]).any().any():
-                    need_insertion.append(obj)
-                # else:
-                    # logging.info("Tuple with github's {} key value already".format(key) +
-                    #     "exists in our db: {}\n".format(str(obj[key])))
-        logging.info("Page recieved has {} tuples, while filtering duplicates this ".format(str(len(new_data))) +
-            "was reduced to {} tuples.\n".format(str(len(need_insertion))))
-        return need_insertion
 
     def assign_tuple_action(self, new_data, table_values, update_col_map, duplicate_col_map, table_pkey):
         """ map objects => { *our db col* : *gh json key*} """
