--- workers/github_worker/tests/test_sample.py
@@ -1,4 +1,0 @@
-# Sample Test passing with nose and pytest
-
-def test_pass():
-    assert True, "dummy sample test"
